only remove the run_script if the RunManager is not running

+++ adaptive_scheduler/server_support.py
@@ -701,7 +701,8 @@ class RunManager:
         from adaptive_scheduler.utils import cleanup_files
 
         with suppress(FileNotFoundError):
-            os.remove(self._default_run_script_name)
+            if self.status() != "running":
+                os.remove(self._default_run_script_name)
 
         return cleanup_files(self.job_names)
